Agent 0.9.0: review fixes (control-user ACL, health token, guard restart, NVML lock)
- install: allow_control_users only get the service control right, no
  longer Modify on the config directory (identity key, provisioning with
  the MQTT password); an old grant is removed on apply-rules. --config is
  stored absolute in the SCM.
- health: POST /restart-child requires X-Agent-Token (control.token next
  to the config). /gpuz stays loopback-only, but the relayed 16-pin values
  no longer drive the guard: high load is judged on NVML board power.
- guard: after a service restart the engine recognises its own targets,
  so a card left at stage 2 is raised back to the permanent limit.
- gpu: the NVML->nvidia-smi fallback and all readers go through the
  mutex; NVML "current" is the management limit, not the enforced one.
- app: a module panic ends the service with an error instead of leaving
  a healthy-looking zombie; provisioning is read under the lock.
- updater: parked .old-<ts> files are cleaned too and the relay restart
  runs in that case; only gpuz-relay processes are ended, not every
  invocation of the binary.
- heartbeat: the /healthz OK window follows the provisioned interval.
- config: identity/logs/tls default next to the loaded --config, not the
  default path; secret_store.domain has no placeholder default and is
  required with password_secret; ollama_proxy needs router.token.
- systemd unit: /usr/local/bin writable so self-update works under
  ProtectSystem=full.
- example config defaults to LocalSystem and documents the missing keys;
  README and design note updated; tests for guard restart, board-power
  load, cleanup glob and config validation.

+++ agent-go/gpuz_windows.go
@@ -158,8 +158,8 @@ func gpuzCandidates() []string {
 	if rc, _, _ := pProcIdToSession.Call(uintptr(windows.GetCurrentProcessId()), uintptr(unsafe.Pointer(&own))); rc == 0 {
 		own = 0xFFFFFFFF
 	}
-	if act, _, _ := pActiveConsoleSess.Call(); act != 0xFFFFFFFF && uint32(act) != own {
-		out = append(out, fmt.Sprintf(`\Sessions\%d\BaseNamedObjects\GPUZShMem`, act))
+	if act, _, _ := pActiveConsoleSess.Call(); uint32(act) != 0xFFFFFFFF && uint32(act) != own { // DWORD: obere 32 Bit nicht definiert
+		out = append(out, fmt.Sprintf(`\Sessions\%d\BaseNamedObjects\GPUZShMem`, uint32(act)))
 	}
 	out = append(out, `\BaseNamedObjects\GPUZShMem`)
 	return out
